% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{Covid19ComplicationsRasInhibitors}
\alias{Covid19ComplicationsRasInhibitors}
\title{Covid19ComplicationsRasInhibitors}
\description{
An OHDSI R package to compare the relative risk of COVID-19 outcomes for RAS inhibitor users
}
